rag/requirements.txt
chromadb==0.4.20
openai==1.51.0
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0 